backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
gTTS
pydub